+++ iMusic/LibraryView/LibraryViewModel.swift
@@ -1,5 +1,5 @@
 //
-//  Library.swift
+//  LibraryViewModel.swift
 //  iMusic
 //
 //  Created by wsa-024-23b on 17.09.21.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-final class Library: ObservableObject {
+final class LibraryViewModel: ObservableObject {
     
     @Published var items: [TrackViewModel] = []
 }
