--- iMusic/Services/LoadingView.swift
@@ -1,27 +1,0 @@
-//
-//  LoadingView.swift
-//  iMusic
-//
-//  Created by wsa-024-23b on 17.09.21.
-//
-
-import SwiftUI
-
-struct LoadingView: View {
-    var body: some View {
-        ZStack {
-            Color(.systemBackground)
-                .ignoresSafeArea()
-            
-            ProgressView()
-                .progressViewStyle(CircularProgressViewStyle(tint: Color("brandColor")))
-                .scaleEffect(2)
-        }
-    }
-}
-
-struct LoadingView_Previews: PreviewProvider {
-    static var previews: some View {
-        LoadingView()
-    }
-}
